Add TwitterSignInCard component
