+++ src/js/routes/Values.jsx
@@ -10,17 +10,8 @@ import PublicFiguresToFollowPreview from '../components/Values/PublicFiguresToFo
 import ValuesFollowedPreview from '../components/Values/ValuesFollowedPreview';
 import ValuesToFollowPreview from '../components/Values/ValuesToFollowPreview';
 import NetworkOpinionsFollowed from '../components/Network/NetworkOpinionsFollowed';
-import TwitterSignIn from '../components/Twitter/TwitterSignIn';
 import VoterStore from '../stores/VoterStore';
-
-const twitterInfoText = (
-  <span className="social-btn-description">
-    <i className="fa fa-info-circle" />
-      Signing into Twitter is the fastest way to find voter guides related to your
-      values. We Vote will find the voter guides for everyone you are following on
-      Twitter
-  </span>
-);
+import TwitterSignInCard from "../components/Twitter/TwitterSignInCard";
 
 // const facebookInfoText = "By signing into Facebook here, you can choose which friends you want to talk politics with, and avoid the trolls (or that guy from work who rambles on)! You control who is in your We Vote network.";
 
@@ -89,21 +80,7 @@ export default class Values extends Component {
             {organizationsBlockToDisplay}
           </div>
 
-          <div className="col-md-4 d-none d-sm-block">
-            {this.state.voter.signed_in_twitter ? null : (
-              <div className="card">
-                <div className="card-main">
-                  <div className="network-btn">
-                    <TwitterSignIn
-                      className="btn btn-social btn-lg btn-twitter btn-twitter-values text-center"
-                      buttonText="Sign In to Find Voter Guides"
-                    />
-                    {twitterInfoText}
-                  </div>
-                </div>
-              </div>
-            )}
-          </div>
+          <TwitterSignInCard voter={this.state.voter} />
         </div>
       </span>
     );
