+++ src/sass/components/_network.scss
@@ -19,19 +19,51 @@ $backdrop-border-radius: 8px;
   color: $gray-mid;
 }
 
-@media (min-width: 768px) {
-  .btn-twitter-values {
-    font-size: 16px;
-    max-width: $button-div-max-width;
+.btn-twitter-values {
+  max-width: $button-div-max-width;
+  width: 100%;
+  .fa-twitter {
+    font-size: inherit;
+    margin-top: -3px;
   }
-  .network-btn {
-    display: inline-block;
+  @include breakpoints (medium) {
+    font-size: 12px;
+    padding: 8px 8px 8px 52px !important;
+  }
+  @include breakpoints (large) {
+    font-size: 14px;
+    padding: 8px 16px 8px 64px;
+  }
+}
+
+.network-btn {
+  display: inline-block;
+  @include breakpoints (small medium) {
+    margin-left: $button-margin-horizontal / 2;
+    margin-right: $button-margin-horizontal / 2;
+  }
+  @include breakpoints (medium) {
     max-width: $button-div-max-width;
     margin-right: 0;
     margin-left: 0;
   }
+}
+
+@media (min-width: 768px) {
   .mobile-container {
-    display: none;
+    @include breakpoints (medium) {
+      display: none;
+    }
+    @include breakpoints (small medium) {
+      display: flex;
+      flex-flow: row nowrap;
+      justify-content: space-around;
+      .btn-social {
+        width: 100%;
+        padding-left: $social-left-padding-mobile;
+        padding-right: $social-right-padding-mobile;
+      }
+    }
   }
 }
 
@@ -40,20 +72,8 @@ $backdrop-border-radius: 8px;
     display: block;
     width: 100%;
   }
-  .network-btn {
-    display: inline-block;
-    margin-left: $button-margin-horizontal / 2;
-    margin-right: $button-margin-horizontal / 2;
-  }
   .mobile-container {
-    display: flex;
-    flex-flow: row nowrap;
-    justify-content: space-around;
-    .btn-social {
-      width: 100%;
-      padding-left: $social-left-padding-mobile;
-      padding-right: $social-right-padding-mobile;
-    }
+    
   }
 }
